feat: support for rendering multiple examples in markdown
