+++ src/utils/parseHTML.ts
@@ -9,21 +9,13 @@ export default function parseHTML(source: string) {
     return null;
   }
 
-  const exampleCode = source.match(/<!-+\ ?start-code\ ?-+>([\s\S]+)<!-+\ ?end-code\ ?-+>/gi);
+  const fragments = source.split(/(<!-+\ ?start-code\ ?-+>[\s\S]+?<!-+\ ?end-code\ ?-+>)/gi);
 
-  if (!exampleCode) {
-    return {
-      beforeHTML: source
-    };
-  }
+  return fragments.map((fragment, key) => {
+    if (fragment.match(/<!-+\ ?start-code\ ?-+>[\s\S]+?<!-+\ ?end-code\ ?-+>/gi)) {
+      return { key, type: 'code', content: text(parseDom(fragment)) };
+    }
 
-  const code = text(parseDom(exampleCode.join('').replace(/\n+/, '')));
-  const beforeHTML = source.match(/([\s\S]+)<!-+\ ?start-code\ ?-+>/gi)?.join('');
-  const afterHTML = source.match(/<!-+\ ?end-code\ ?-+>([\s\S]+)/gi)?.join('');
-
-  return {
-    code,
-    beforeHTML,
-    afterHTML
-  };
+    return { key, type: 'html', content: fragment };
+  });
 }
